# Conflicts:
#	src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/generator/LargeCombustionEngineMachine.java

+++ src/main/java/com/gregtechceu/gtceu/common/cover/data/BucketMode.java
@@ -17,9 +17,9 @@ public enum BucketMode implements EnumSelectorWidget.SelectableEnum {
     @Getter
     public final IGuiTexture icon;
 
-    public final long multiplier;
+    public final int multiplier;
 
-    BucketMode(String tooltip, String textureName, long multiplier) {
+    BucketMode(String tooltip, String textureName, int multiplier) {
         this.tooltip = tooltip;
         this.icon = new ResourceTexture(textureName + ".png").scale(16F / 20F);
         this.multiplier = multiplier;
